+++ plugins/typings/test/typings.test.js
@@ -1,0 +1,5 @@
+const typingsPlugin = require('../src').default
+
+describe('typings', () => {
+  
+})
